backend/requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_migrate==4.0.5
flask_cors==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
